Split Azure DevOps oversized tests
